CLI as workspace root dep
