+++ package.json
@@ -40,6 +40,7 @@
   "devDependencies": {
     "@ark/attest": "0.11.0",
     "@changesets/cli": "^2.27.7",
+    "@latticexyz/cli": "workspace:*",
     "@types/node": "^18.15.11",
     "@typescript-eslint/eslint-plugin": "7.1.1",
     "@typescript-eslint/parser": "7.1.1",
